coffee-scraper/requirements.txt
playwright>=1.49.0
asyncpg>=0.30.0
pydantic-settings>=2.5.0
openai>=1.50.0
httpx>=0.27.0
PyPDF2>=3.0.0
python-pptx>=0.6.23
python-dotenv>=1.0.0
cryptography>=42.0.0
google-auth>=2.20.0
